Better debugging; Granular World Environment; And Sequencing tools!
Print uses a better system of displaying metadata of the calling type. This helps to find specifically which instance in the remote tree is causing troubles, if only one is.

SettingsCompliantWorldEnvironment now uses a configuration file backed world environment, so default settings can be more easily set for individual projects when using a symlink approach (as I do)

Sequencing Tools: `SequenceTrigger`s act upon some kind of sensor impulse and perform one or more `SequenceAction`s. Currently the bare-bones system is that the `SequenceActionSpawnable` can spawn any number of packed scenes at the trigger's specified target. The Sequencing tooling allows for implementation of custom triggers and actions so long as proper implementation is respected. See the base class for each for documentation.

Removing the reticle! There are some games that would rather not have a reticle (crosshairs) at all. That is toggleable in the configuration file. Otherwise it behaves as normal.