Formatted GUI to be more space efficient
Originally I believe that changing the theme resource would solve the space efficiency problem. But as fate would have it, I was quite dumb when originally creating these scenes and made a ton of theme overrides instead of just adjusting the theme to make something consistent. So I did my best to also purge theme overrides. Additionally, the project.godot file has the window size set to 1920x1080, which the resolution of my desktop. This should make it easier to tell if a GUI element will be space efficient or not