Implemented API Key Management
Basic security implemented. Keys are stored in a single type, which should be consumed by the API handlers. For example, Discord Rich Presence, which I hope to implement given it's easy and works well

In order to implement this feature, some more packages were needed. Refer to the changes in SquigglesCommon4X.csproj for details